+++ app/src/main/java/com/gdgssu/android_deviewsched/ui/MainActivity.java
@@ -67,7 +67,7 @@ public class MainActivity extends AppCompatActivity implements DeviewFragment.On
                 actionBar.setDisplayShowTitleEnabled(true);
                 actionBar.setDisplayUseLogoEnabled(false);
                 actionBar.setHomeButtonEnabled(true);
-                actionBar.setHomeAsUpIndicator(R.drawable.ic_menu_black_24dp);
+                actionBar.setHomeAsUpIndicator(R.drawable.ic_menu_white_24dp);
                 //흰색으로 바꿀까 말까 고민중...
             }
         }
