+++ app/src/main/java/com/gdgssu/android_deviewsched/ui/sche/ScheFragment.java
@@ -73,7 +73,13 @@ public class ScheFragment extends DeviewFragment {
         ((AppCompatActivity)getActivity()).setSupportActionBar(mToolbar);
 
         ActionBar actionBar = ((AppCompatActivity)getActivity()).getSupportActionBar();
-        actionBar.setHomeAsUpIndicator(R.drawable.ic_menu_black_24dp);
         actionBar.setDisplayHomeAsUpEnabled(true);
+
+        mToolbar.setNavigationOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                ((MainActivity)getActivity()).showHome();
+            }
+        });
     }
 }
