+++ app/src/main/res/values/styles.xml
@@ -28,7 +28,6 @@
     <style name="DeviewAppTheme_MaterialViewPager" parent="DeviewAppBaseTheme_MaterialViewPager">
 
         <item name="android:textColorPrimary">@android:color/black</item>
-        <item name="drawerArrowStyle">@style/DrawerArrowStyle</item>
         <item name="android:windowTranslucentStatus" tools:targetApi="21">true</item>
 
         <item name="android:windowContentOverlay">@null</item>
@@ -37,7 +36,6 @@
 
         <!-- Toolbar Theme / Apply white arrow -->
         <item name="colorControlNormal">@android:color/white</item>
-        <item name="actionBarTheme">@style/AppTheme.ActionBarTheme</item>
 
         <!-- Material Theme -->
         <item name="colorPrimary">@color/colorPrimary</item>
@@ -47,16 +45,6 @@
         <item name="android:navigationBarColor" tools:targetApi="21">@color/navigationBarColor</item>
         <item name="android:windowDrawsSystemBarBackgrounds" tools:targetApi="21">true</item>
 
-    </style>
-
-    <style name="AppTheme.ActionBarTheme" parent="@style/ThemeOverlay.AppCompat.ActionBar">
-        <!-- White arrow -->
-        <item name="colorControlNormal">@android:color/white</item>
-    </style>
-
-    <style name="DrawerArrowStyle" parent="Widget.AppCompat.DrawerArrowToggle">
-        <item name="spinBars">true</item>
-        <item name="color">@color/drawerArrowColor</item>
     </style>
